+++ Framework/Database/PermissionsCenterDB/SC/Types/ObjectWithParentTable.sql
@@ -5,7 +5,7 @@ CREATE TYPE [SC].[ObjectWithParentTable] AS TABLE
 	SchemaType NVARCHAR(64) NULL,
 	Name NVARCHAR(64) NULL,
 	ParentID NVARCHAR(36) NOT NULL,
-	InnerSort INT NULL,
+	InnerSort BIGINT NULL,
 	FullPath NVARCHAR(414) NULL,
 	GlobalSort NVARCHAR(414) NULL --,
 	--PRIMARY KEY ([ID] ASC, [ParentID] ASC)
